refactor: tokencontroller 삭제 및 로직 service로 이전
