+++ src/main/java/com/example/gocalowly/domain/token/service/TokenService.java
@@ -1,21 +1,20 @@
-package com.example.gocalowly.domain.token.controller;
+package com.example.gocalowly.domain.token.service;
 
-import com.example.gocalowly.domain.token.service.AccessTokenService;
-import com.example.gocalowly.domain.token.service.RefreshTokenService;
 import jakarta.servlet.http.HttpServletResponse;
-import java.util.UUID;
-import org.springframework.stereotype.Controller;
+import org.springframework.stereotype.Service;
 
-@Controller
-public class TokenController {
+import java.util.UUID;
+
+@Service
+public class TokenService {
+
     AccessTokenService accessTokenService;
     RefreshTokenService refreshTokenService;
 
-    public TokenController(AccessTokenService accessTokenService, RefreshTokenService refreshTokenService) {
+    public TokenService(AccessTokenService accessTokenService, RefreshTokenService refreshTokenService){
         this.accessTokenService = accessTokenService;
         this.refreshTokenService = refreshTokenService;
     }
-
 
     public void setTokens(UUID userId, HttpServletResponse response) {
         String accessToken = accessTokenService.generateAccessToken(userId);
